fix: 로또 수익률 계산에서 아주 큰수 혹은 아주 작은 수가 들어왔을 때 예상치 못한 결과가 나오는것을 방어함

+++ src/main/java/lotto/view/OutputView.java
@@ -1,5 +1,6 @@
 package lotto.view;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -53,7 +54,7 @@ public class OutputView {
     /**
      * 백분위 수익률 출력
      */
-    public static void printPrizePercentResult(double out) {
+    public static void printPrizePercentResult(BigDecimal out) {
         System.out.print("총 수익률은 " + out + "%" + "입니다.");
     }
